requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
colorio==0.7.3
    # via -r requirements.in
colormath==3.0.0
    # via -r requirements.in
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via networkx
font-fredoka-one==0.0.4
    # via inky
font-hanken-grotesk==0.0.2
    # via inky
font-intuitive==0.0.4
    # via inky
font-source-serif-pro==0.0.1
    # via inky
git+https://www.github.com/hbldh/hitherdither
    # via -r requirements.in
importlib-metadata==4.5.0
    # via colorio
inky[fonts,rpi]==1.2.0
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.4.2
    # via colorio
networkx==2.5.1
    # via colormath
numpy==1.20.3
    # via
    #   -r requirements.in
    #   colorio
    #   colormath
    #   hitherdither
    #   inky
    #   matplotlib
pillow==8.2.0
    # via
    #   -r requirements.in
    #   hitherdither
    #   matplotlib
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via matplotlib
rpi.gpio==0.7.0
    # via inky
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
smbus2==0.4.1
    # via inky
spidev==3.5
    # via inky
typing-extensions==3.10.0.0
    # via importlib-metadata
zipp==3.4.1
    # via importlib-metadata
